requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langgraph>=0.1.0
faiss-cpu>=1.7.4
openai>=1.30.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
google-search-results>=2.4.2   # SerpAPI Python SDK
streamlit>=1.35.0              # optional – for the web UI
